Homework_10 remaked and added test_loger.py
